add support for java, secure endpoints

+++ scripts/runTest.sh
@@ -10,7 +10,15 @@ SOURCE_PATH="submissions/$SUBMISSION_ID"
 cp "$DATA_PATH/$DATASET_ID/in" "$SOURCE_PATH"
 cp "$DATA_PATH/$DATASET_ID/out" "$SOURCE_PATH"
 
-docker run --rm -v $(pwd)/$SOURCE_PATH:/judge -w /judge debian /bin/sh -c "timeout $TIME_LIMIT ./$SUBMISSION_ID < in > myout"
+if [ "$5" == "java" ]; then
+    filename=$6
+    filename=${filename%.java}
+    pushd $SOURCE_PATH
+    timeout $TIME_LIMIT java $filename < in > myout
+    popd
+else
+    docker run --rm -v $(pwd)/$SOURCE_PATH:/judge -w /judge debian /bin/sh -c "timeout $TIME_LIMIT ./$SUBMISSION_ID < in > myout"
+fi
 EXIT_CODE=$?
 
 if [ "$EXIT_CODE" != "0" ]; then
